renamed mentions of thapbi_santi to pycits
This also involves changes to setup.py (and a corresponding move
of the main script to the bin/ subdirectory to avoid post-install
module import problems with distutils), and some local/relative
import changes in the pycits module files.
