+++ pycits/__init__.py
@@ -1,5 +1,3 @@
-import fastqc
-
 # python package version
 # should match r"^__version__ = '(?P<version>[^']+)'$" for setup.py
-__version__ = '0.1.0'
+__version__ = '0.1.0.dev'
