+++ pycits/blast.py
@@ -9,7 +9,7 @@ import os
 import sys
 
 from subprocess import Popen, PIPE
-from tools import is_exe
+from .tools import is_exe
 
 
 class Blastclust(object):
