+++ pycits/ea_utils.py
@@ -11,4 +11,4 @@ import os
 import sys
 
 from subprocess import call
-from tools import is_exe
+from .tools import is_exe
