+++ pycits/muscle.py
@@ -9,7 +9,7 @@ import os
 import sys
 
 from subprocess import call
-from tools import is_exe
+from .tools import is_exe
 
 
 class Muscle(object):
