+++ pycits/qiime.py
@@ -11,7 +11,7 @@ import os
 import sys
 
 from subprocess import call
-from tools import is_exe
+from .tools import is_exe
 
 
 class Pick_Otus(object):
